+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Added tracing to the `miden-tx-prover` CLI (#1014).
 - Added metrics to the `miden-tx-prover` proxy (#1017).
 - Implemented `to_hex` for `AccountIdPrefix` and `epoch_block_num` for `BlockHeader` (#1039).
+- Introduce `AccountIdBuilder` to simplify `AccountId` generation in tests (#1045).
 
 ## 0.6.2 (2024-11-20)